requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
pymupdf==1.23.8
pillow==10.1.0
python-socketio==5.10.0